+++ src/app/[elementType]/[elementID]/[[...tab]]/_IcreTabs/_Variants1/GWASLdr.tsx
